Started implementing from_ndds classmethod
